+++ C++ Codes/bintodec.cpp
