+++ C++ Codes/factor.cpp
