+++ C++ Codes/prime.cpp
