+++ Java Codes/YoungPhysicist.java
@@ -1,4 +1,3 @@
-package codeforces;
 import java.util.Scanner;
 
 public class YoungPhysicist {
