+++ Java Codes/gcdOfTwoNumbers.java
